:recycle: Put components in their own modules for re-use with actions
