+++ src/openforms/js/components/admin/form_design/logic/Trigger.js
@@ -2,76 +2,11 @@ import React, {useContext, useState} from 'react';
 import PropTypes from 'prop-types';
 
 import Select from '../../forms/Select';
-import {TextInput, NumberInput} from '../../forms/Inputs';
+
 import {ComponentsContext} from './Context';
-
-
-const ComponentSelection = ({value, onChange}) => {
-    const allComponents = useContext(ComponentsContext);
-    const choices = Object.entries(allComponents).map( ([key, comp]) => [key, comp.label || comp.key] )
-    return (
-        <Select
-            name="trigger.component"
-            choices={choices}
-            allowBlank
-            onChange={onChange}
-            value={value}
-        />
-    );
-};
-
-ComponentSelection.propTypes = {
-    value: PropTypes.string.isRequired,
-    onChange: PropTypes.func.isRequired,
-};
-
-
-// TODO: these are the built in json logic operators, but it's possible to define
-// custom operators. See https://jsonlogic.com/operations.html
-// We start with a minimal supported set based on the user stories (and common sense).
-// The data structure is a mapping of the actual json logic operator as key and the
-// human readable label as value.
-const OPERATORS = {
-    '==': 'is equal to',
-    '!=': 'is not equal to',
-    '>': 'is greater than',
-    '>=': 'is greater than or equal to',
-    '<': 'is less than',
-    '<=': 'is less than or equal to',
-    'in': 'in', // array or string (!)
-};
-
-
-// map the Formio component types to allowed operators. Strings cannot be compared
-// against each other for >=,>,<,<= etc., so this includes a logical allow-list to only
-// show relevant components.
-const COMPONENT_TYPE_TO_OPERATORS = {
-    number: [
-        '==',
-        '!=',
-        '>',
-        '>=',
-        '<',
-        '<=',
-    ],
-    textfield: [
-        '==',
-        '!=',
-        'in',
-    ],
-    iban: [
-        '==',
-        '!=',
-        'in',
-    ],
-};
-
-
-const COMPONENT_TYPE_TO_INPUT_TYPE = {
-    number: NumberInput,
-    textfield: TextInput,
-    iban: TextInput,
-};
+import { OPERATORS, COMPONENT_TYPE_TO_OPERATORS } from './constants';
+import ComponentSelection from './ComponentSelection';
+import LiteralValueInput from './LiteralValueInput';
 
 
 const OperatorSelection = ({selectedComponent, operator, onChange}) => {
@@ -132,51 +67,6 @@ OperandTypeSelection.propTypes = {
 };
 
 
-
-const LiteralValueInput = ({componentType, value='', onChange}) => {
-    const InputComponent = COMPONENT_TYPE_TO_INPUT_TYPE[componentType] || TextInput;
-
-    const onInputChange = (event) => {
-        const inputValue = event.target.value;
-        let value;
-
-        // do any input type conversions if needed, e.g. date to native datetime/ISO-8601 format
-        switch (componentType) {
-            case 'number': {
-                value = Number.parseFloat(inputValue);
-                break;
-            }
-            default:
-                value = inputValue;
-        }
-
-        onChange({
-            target: {
-                name: event.target.name,
-                value: value,
-            }
-        })
-    };
-
-    return (
-        <InputComponent
-            name="trigger.literalValue"
-            value={value}
-            onChange={onInputChange}
-        />
-    );
-};
-
-LiteralValueInput.propTypes = {
-    componentType: PropTypes.string.isRequired,
-    value: PropTypes.oneOfType([
-        PropTypes.string,
-        PropTypes.number,
-    ]),
-    onChange: PropTypes.func.isRequired,
-};
-
-
 const Trigger = ({ id }) => {
     const [triggerComponent, setTriggerComponent] = useState('');
     const [operator, setOperator] = useState('');
@@ -193,6 +83,7 @@ const Trigger = ({ id }) => {
         case 'literal': {
             valueInput = (
                 <LiteralValueInput
+                    name="trigger.literalValue"
                     componentType={componentType}
                     value={literalValue}
                     onChange={event => {
@@ -228,6 +119,7 @@ const Trigger = ({ id }) => {
 
             <div>
                 <ComponentSelection
+                    name="trigger.component"
                     value={triggerComponent}
                     onChange={event => setTriggerComponent(event.target.value)}
                 />
